requirements.txt
aerich==0.8.2
aiomysql==0.2.0
aiosqlite==0.21.0
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
asyncclick==8.1.8.0
beautifulsoup4==4.13.3
billiard==4.2.1
celery==5.4.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography>=42.0.0
dictdiffer==0.9.0
et_xmlfile==2.0.0
fastapi==0.115.11
h11==0.14.0
httptools==0.6.4
idna==3.10
iso8601==2.1.0
Jinja2==3.1.6
kombu==5.5.0
lxml==5.3.1
MarkupSafe==3.0.2
openpyxl==3.1.5
prompt_toolkit==3.0.50
pydantic==2.9.0
pydantic-settings==2.8.1
pydantic_core==2.23.2
PyMySQL==1.1.1
pypika-tortoise==0.5.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2025.1
PyYAML==6.0.2
redis==5.2.1
requests==2.32.3
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
starlette==0.46.1
tomlkit==0.13.2
tortoise-orm==0.24.2
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
vine==5.1.0
watchfiles==1.0.4
wcwidth==0.2.13
websockets==15.0.1
